Migrate from molecule to packer entirely
molecule is being deprecated and our build process is being changed to
support using infrastructure available within twitch. These changes are
made such that builds can still take place outside of twitch. To this
end there are two primary additions.

 - ./build.sh
 - ./build/packer.json

These work together to prep the environment for packer to build an AMI.
Packer now replaces much of what molecule was doing. `build.sh` handles
the pre-packer stages of the build including building the binary, which
it cross compiles for linux/amd64.

Note: the build is not responsible for running godep restore and test,
that falls on the shoulders of the person(/thing) running `./build.sh`