add migration endpoint
Summary:
The ingester needs to know how to migrate from one version of the schema to
another. This adds an endpoint to blueprint to expose the answer to that
question. The ingester will eventually hit this and convert it into SQL.

Test Plan:
Added some basic tests around the bad arguments. More interesting tests will
come with the ingester work.

Reviewers: maasted

Reviewed By: maasted

Differential Revision: https://phabricator.internal.justin.tv/D3570