blueprint: kill scoop, cuts over to only bp migrations, validates migrations
Summary:
This removes the scoop dependency from blueprint, using only the bpdb schema
store for migrations. This also adds validation of the schemas to blueprint
itself, instead of relying on ALTERs failing through scoop.

This will be merged with the ingester migrator change, in some smart order that
I haven't yet thought through.

Test Plan:
go test ./... passes
will be doing some kind of testing with this AND the ingester change, TBD

Reviewers: maasted

Reviewed By: maasted

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D3709