blueprint refactor: use action metadata for operations
Summary:
For the rename column feature on blueprint, we need to add the concept of
a rename to the migration states. Previously we had only `"add"` and `"delete"`,
corresponding to adding a column and deleting a column. The schema for bpdb's
operation table worked well for `add`, as it held all the requisite information
for adding a column, and `delete` was fine too, as the only thing that operation
really needed to store was the outbound name of the column that was being
deleted.

For the rename operation, we need to have both the old outbound name and the new
outbound name. That doesn't fit well into the previous schema, so Jack, Matt
and I brainstormed a bit and decided on the schema here: add a jsonb column for
any special data a particular operation needs, drop all the columns that don't
apply to all operations, and put that data into the jsonb field.

This diff is the blueprint side of the refactor work without adding the rename
feature. Most of the diff is that, with some extra refactoring that's not
necessarily tied to that change.

  - There is a diff for scoop_protocol: https://phabricator.internal.justin.tv/D4118
  - There is a diff for the ingester too to accept the new `/migration` format: https://phabricator.internal.justin.tv/D4121
  - There is a diff for the integration test to use the new format: https://phabricator.internal.justin.tv/D4122
  - This is the SQL I'm planning on running to migrate the db. I ran it on my integration db before the int test with no problems: https://git-aws.internal.justin.tv/mperrone/scratch/blob/master/blueprint_testing/migrate_bp.sql

Test Plan:
`go test ./...` passes, and I used the integration environment to test the
important endpoints `/schema`, `/schemas`, `/migration` manually.

Integration test passes.

Deploy Plan:
* Land the scoop_protocol change.
* Update the ingester and blueprint diffs to include the landed scoop_protocol change.
* Land blueprint
* Land ingester
* Turn off the rsloadmanager (leave metadatastorer on)
* Turn off blueprint
* Migrate blueprint db
* Deploy new blueprint - verify
* Deploy new ingester

Reviewers: jackgao, ojcm

Reviewed By: ojcm

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D4119