Add better read-only mode
Summary:
Make non-logged-in users able to use the UI in a read-only manner.
Make users explicitly log in to make schema changes.
Disable/remove inputs on schema list and schema creation (to support viewing schema suggestions).
Leave the schema editing page as-is because it's not linked to.
Login checking is done via a cookie since the server is performing access control.
Remove notes/suggestions on the schema creation page since they are unused.
Fix a bug in renamed outbound columns disappearing on schema update.
Redirect the root page over http to https in nginx.
Prevent saving when a column is partially added.

Test Plan: Tested UI and interactions locally logged in and logged out.

Reviewers: mperrone

Reviewed By: mperrone

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D4281