check for dupe columns in the UI and backend
Summary:
Blueprint allowed creating duplicate column names, which would cause errors
later when trying to build schemas for /schemas. This changes the UI and the
backend to reject duplicate column names.

from:
https://twitchtv.atlassian.net/browse/SCIENG-506

Test Plan:
ran it locally and tested both using the UI to create a schema with dupe
columns, and CURL'd a request to the backend to create a schema with dupe
columns, both were rejected.

integration test still passes

Reviewers: ojcm

Reviewed By: ojcm

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D4353