Add ability to drop tables and request them to be dropped.
Summary:
Allow tables that don't exist yet in Redshift to be dropped.
Any other tables can have a drop requested, which will notify
the team via slack and mark them as being dropped in the UI.
Once tables are dropped, they are no longer visible on blueprint,
and they can be recreated.
For complicated reasons, a table can't be automatically dropped
even if it doesn't exist on Redshift if it's already been dropped
and recreated.

Test Plan: Ran integration test.

Reviewers: mperrone

Reviewed By: mperrone

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D4445