Add cache of all schemas result to Blueprint
Summary:
Selecting all schemas from the Blueprint database takes about 1 s, which is
too slow for replay (which queries one for each instance of the processor it
spawns, which is a lot).  Since the result only changes when a schema is
created or updated, this is a prime target for caching.

I ended up not using the external in-memory cache, as it didn't have quite the
characteristics I was looking for.

Test Plan: Unit and integration tests pass.

Reviewers: jackgao

Reviewed By: jackgao

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D4448