Adding support for mapping transformers
Summary:
In preparation for name changes being made into a public feature, this commit brings support for the new the transformer type userIDWithMapping. This transformer sets the value of a column expected to be a user id when is empty/invalid, and for that it expects a supporting column with a login so we can use it to map to an id through an underlying API.

On the Blueprint side, the following was required:

- Upgraded part of scoop_protocol to the most recent version with the new ColumnDefinition and transformer name.
- The Blueprint backend was updated so all Add operations now require supporting columns as part of the metadata. For other transformers is expected to be empty, but for the new transformer is expected to be a comma separated list of column names.
- UI changes were made so users can now provide supporting columns when adding a columns that is of the new transformer type. Validation is also in place to make sure only valid columns are used as part of the list.
- When suggesting columns for an event, by default we suggest that user_id, channel_id and host_channel_id should use the new transformer type with the expected supporting column by convention.

Additionally, the following are links to required changes in other projects:

- [[ https://phabricator.internal.justin.tv/D4779 | rs_ingester ]]
- [[ https://phabricator.internal.justin.tv/D4780 | scieng-deploy ]]

Test Plan: Ran successfully a modified version of integration that sends events with supporting columns and using the new transformer. I also connected to the integration blueprint and tested adding valid/invalid columns with/without the new transformer and verified it reacted as expected. For production, I intend to test the blueprint website and make sure listing/creating/updating works with a dummy table.

Reviewers: dwe

Reviewed By: dwe

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D4778