Fixing the inbound names select on create
Summary: We need to consider duplicate inbound names. Also added a check to make sure the selected inbound is valid just in case.

Test Plan: Verified on the integration page that inbound names are not being duplicated and a table can be created. On prod, I'll create a dummy table to verify the same.

Reviewers: dwe

Reviewed By: dwe

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D4840