Add maintenance mode table
Summary:
This table will track when maintenance mode is turned on or off, as well as
its current state (the query for that will look at the maximum timestamp,
hence the index).

Test Plan:
Ran the integration test, then looked in the Blueprint DB and verified that
the correct table and row was present.

Reviewers: jackgao

Reviewed By: jackgao

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D4839