Build API to maintenance mode settings
Summary:
* Blueprint reads maintenance mode from the DB table on startup.
* Any member of the organization has authorization to view maintenance mode.
* Members of a specified admin team may turn maintenance mode on or off.
* Any operation which modifies the DB first checks the maintenance-mode bit
  and fails if it is set.
* Also refactored the schema generator tests to compile under Go 1.7.

Test Plan:
Temporarily added forms to the UI which set and unset maintenance mode.
Verified that these worked when I was logged in and not when I logged out,
and that modifying schemas failed when maintenance mode was on.  In prod,
will check that Blueprint is still on and that the pipeline is running;
there should be no functionality changes yet, as the UI is still in progress.
(There will be a config change.)

Reviewers: jackgao

Reviewed By: jackgao

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D4889