Add maintenance mode UI to Blueprint
Summary:
* When an member of the designated admin team is logged in, a maintenance
  mode toggle will appear on the front page.  Clicking it brings up a form
  asking for the reason for the change.
* When maintenance mode is on, all pages will appear as if the user is not
  logged in, except for the line at the top "Logged in as <user>" and a banner
  explaining that maintenance mode is on and changes are impossible.
* Unrelated: when Blueprint queries GitHub for organization and team
  membership, it now properly query-escapes the names of the organization
  or team.

Test Plan:
In local testing with dummy auth (the unknown user is now designated an
admin), verified that turning maintenance mode on/off works, but only when a
reason is specified; and that when maintenance mode is on, buttons that allow
modifying the database do not appear.  Also verified that logging out and in
again while in maintenance mode does not break anything.

Will also test locally with real auth and run the integration test.

In production, will manually verify that navigating Blueprint still works
and will turn on maintenance mode briefly to verify it as well (will let the
Science team know that this is happening).

Reviewers: mperrone

Reviewed By: mperrone

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D4908