Fix issue with schema suggestor and mapping columns
Summary:
Schema suggestions with channel/user ids recommend using userIDWithMapping,
but if a user changes it to bigint, it doesn't clear the SupportingColumns.
This causes messy data in blueprint, but the spade transformer is robust, so no
operational issues arose.

Test Plan:
Ran locally and tested a schema suggestion. Ran integration test. Will test on
a schema suggestion in prod.

Reviewers: clgroft

Reviewed By: clgroft

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D5213