Migrating Kinesis Configurations on Blueprint
Summary: We want to start managing Kinesis configurations on Blueprint through a UI. This is the Blueprint piece to that. Before you read the code, you may want to play around with it on my integration Blueprint. I added a set of API that are admin-only, and their corresponding logic that leads to editing a database in bpdb `kinesis_config`. The _vendor mess is mainly because I introduced a dependency to processor, where we use the kinesis writer configurations from there. I'd be happy to walk you though this.

Test Plan: Tested on integration environment with some real configurations from processor.

Reviewers: clgroft

Reviewed By: clgroft

Subscribers: maasted, #scieng

Differential Revision: https://phabricator.internal.justin.tv/D5189