enforce time must be present in all schemas and enforce that date cannot be
Summary:
This is to pave the way for having data in spectrum. We want the partition key
to be called date, which is effectively a column just in spectrum, so we don't
want a collision. Date will be generated from time, so we want to be very sure
that it stays there.

This enforces that at the ui and in the api.

I'd like help on how to have the blacklist in angular in the two different
files. Right now I just copy and pasted since I'm not sure how to do better.

Test Plan:
- undid the commits for the ui changes to test the api: I validated that I get
  errors for not having time on a new one, and renaming time away
- tried them in the ui and got the ui errors
- go test ./... for the api changes.
- tests for the UI seem to not exist, and I really don't want to blaze that
  trail in angular :|

Reviewers: jackgao

Reviewed By: jackgao

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D5371