Update invalid schema url handling
Summary:
This change is to align the behavior of the event comment and event metadata endpoints with the behaviour of `/schema` and `/droppable` when the schema does not exist (return 404 rather than 500). This means event comment errors from the schema not existing will no longer send errors to Rollbar, but just 404 instead. The user will still see be redirected to the landing page and see an error message `API Error: Not Found`.

See https://twitchtv.atlassian.net/browse/SCIENG-1460 for more details.

- Update event comment and event metadata handlers to return 404 when schema not found
- Update handler tests and mock backend

Test Plan:
- `go test ./... passes`
- Integration builds successfully
- Changing the url to a schema that does not exist `/#/schema/invalid-table` leads to the landing page with an error message `API Error: Not Found` from both the landing page and a specific schema page
	- No 500 errors in Network tab

Reviewers: mperrone, lydon

Reviewed By: mperrone, lydon

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D5402