Update scoop_protocol to remove stringToIntegerMD5
Summary:
stringToIntegerMD5 was originally intended to hash user agents for privacy, and
possibly strings to use as distkeys. It is unused and the ingester doesn't
know it exists, so it's safer to destroy it.

Test Plan:
Unit and integration tests pass.  Also checked in browser that
stringToIntegerMD5 is no longer an option for column type.

Reviewers: dwe

Reviewed By: dwe

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D5421