Add ID and region fields to database
Summary:
To support writing to streams in other regions, add a StreamRegion field to Kinesis configs.
Soon, we'll be moving to using an ID as a primary column to simplify lookups,
so add an ID column that is populated correctly but not used for anything.
We currently can't have two streams in the same account in different regions, but that's ok
for now.

Deploy Plan:
1. Run the init/migrate.sql script.
2. Deploy the new version of blueprint.
3. Validate streams are being written to correctly and that updates work as expected.
4. Tear down the old version of blueprint.
5. Run the last section of the init/migrate.sql script again.

Test Plan: Ran integration test. Modified/dropped/created kinesis configs in integration.

Reviewers: jackgao

Reviewed By: jackgao

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D5494