per-schema maintenance mode
Summary:
It would not be good if a schema migration happened in the middle of UNLOADing data for a tahoe backfill.
This allows for that by adding per table schema maintenance mode. Adds the
buttons to the admin ui, endpoints for reading and writing, and a new table to
store it. Follows the conventions of the global maintenance mode mostly.

In most places, maintenance mode has been renamed to global maintenance mode to
differentiate with schema maintenance mode.

Out of spec it adds reporting the user that turned maintenance mode on for global, as
well as some minor refactoring of those code paths.

Rollout Plan:
```
ALTER TABLE maintenance ADD "user" text;
    UPDATE maintenance SET "user" = 'legacy';
    ALTER TABLE maintenance RENAME TO global_maintenance;
    ALTER INDEX maintenance_ts_index RENAME TO global_maintenance_ts_index;

    CREATE TABLE schema_maintenance
    (
      ts timestamp without time zone default NOW(),
      schema text,
      is_maintenance boolean,
      "user" text,
      reason varchar
    );
    CREATE INDEX schema_maintenance_ts_index ON schema_maintenance(ts);
```
(ran these on integration before testing the init.sql bootstrapping)

Test Plan:
- integration tests passes
- `go test ./...` passes
- on integration validated that:
  - startup picks up the current maintenance modes
  - new tables don't fail on `/maintenance`
  - setting on and off maintenance mode global+one schema disables the ui (on
    reload for global)

Reviewers: clgroft

Reviewed By: clgroft

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D5499