Make auth network calls stubbable for testing
Summary:
The calls that authentication makes to the GitHub server are now
handled by a subobject which is stubbed out for testing.  Moved the
existing auth_test.go (which is never run AFAICT) to
google_auth_test.go (I'd be open to just deleting it), and added a new
test using the stub to auth_test.go.  Will add more tests when I have
approval for the general method.

Test Plan: The new tests work.

Reviewers: lydon

Reviewed By: lydon

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D5536