Running angular tests on arc diff
Summary: Creating a custom php class that extends from ArcanistUnitTestEngine which just executes run_angular_tests.sh whenever arc diff is invoked, preventing the diff creation when it fails and prints the output for debugging. An .arcconfig file had to be created to point to this new custom class.

Test Plan: I tested by running `arc lint` with and without a broken angular file and verfied id did the right thing. Also, if this diff was created it means it didn't block a perfectly fine diff.

Reviewers: dwe

Reviewed By: dwe

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D5568