Adding tests for blueprint.kinesisconfig.create
Summary: Adding test to verify that we can create kinesis configurations when the right conditions are met and fails in expected ways when not. As part of the changes, discovered some fixes and clean up that could be done in create.js

Test Plan: Ran new tests. Also tested on my integration the different kinesis creation paths. On prod will make sure things are healthy.

Reviewers: lydon

Reviewed By: lydon

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D5648